Improve the build environment
In `Release` builds, constant TRACE is not wanted. Replace it with RELEASE.

References do not specify assembly version, public key, etc. Add that detail.

The post-build script contains hard-coded paths and is somewhat opaque as to
how the local C:SL folders are structured. Make use of macros and environment
variables to generalise the script, and echo folder paths to build console as
visual confirmation that the correct paths are being used.